# backend/solarhub-nestjs/swagger-output.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/offers/available-requests": {
      "get": {
        "operationId": "CompanyOffersController_listAvailableRequests",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Offers"
        ]
      }
    },
    "/api/v1/offers/requests/{id}/reply": {
      "post": {
        "operationId": "CompanyOffersController_replyToRequest",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Offers"
        ]
      }
    },
    "/api/v1/offers/my-offers": {
      "get": {
        "operationId": "CompanyOffersController_listCompanyOffers",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Offers"
        ]
      }
    },
    "/api/v1/offers/my-offers/{id}": {
      "get": {
        "operationId": "CompanyOffersController_getCompanyOffer",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Offers"
        ]
      },
      "put": {
        "operationId": "CompanyOffersController_updateCompanyOffer",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Offers"
        ]
      },
      "delete": {
        "operationId": "CompanyOffersController_deleteCompanyOffer",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Offers"
        ]
      }
    },
    "/api/v1/offers/my-offers/{id}/finish": {
      "post": {
        "operationId": "CompanyOffersController_finishCompanyOffer",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Offers"
        ]
      }
    },
    "/api/v1/offers/my-offers/{id}/complete": {
      "post": {
        "operationId": "CompanyOffersController_completeCompanyOffer",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Offers"
        ]
      }
    },
    "/api/v1/offers/involvements": {
      "get": {
        "operationId": "InvolvementTemplatesController_listInvolves",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Involvement Templates"
        ]
      },
      "post": {
        "operationId": "InvolvementTemplatesController_createInvolve",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvolvementTemplateInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Involvement Templates"
        ]
      }
    },
    "/api/v1/offers/involvements/{id}": {
      "put": {
        "operationId": "InvolvementTemplatesController_updateInvolve",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvolvementTemplateInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Involvement Templates"
        ]
      },
      "delete": {
        "operationId": "InvolvementTemplatesController_deleteInvolve",
        "parameters": [
          {
            "name": "company-id",
            "in": "header",
            "description": "Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Involvement Templates"
        ]
      }
    },
...
  },
  "info": {
    "title": "SolarHub API",
    "description": "The SolarHub Backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "OfferInDto": {
        "type": "object",
        "properties": {}
      },
      "InvolvementTemplateInDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}